requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==24.3.0
av==14.0.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.5.0
faster-whisper==1.1.0
filelock==3.16.1
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
multidict==6.1.0
numpy==2.2.1
onnxruntime==1.20.1
packaging==24.2
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pvrecorder==1.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.10.4
pydantic_core==2.27.2
pyht==0.1.10
pyparsing==3.2.0
pyreadline3==3.5.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==13.1
yarl==1.18.3
